+++ HISTORY.rst
@@ -2,7 +2,8 @@
 History
 =======
 
-0.1.0 (2016-12-22)
+0.8.0 (2017-03-16)
 ------------------
 
-* First release on PyPI.
+* ``conda-devenv`` now can receive standard ``environment.yml`` files, in which case the file
+  will just be forwarded to ``conda env update`` normally.